Add raw data provider test case, enable saving rotated data
